+++ epistasis/models/nonlinear.py
@@ -142,17 +142,6 @@ class EpistasisNonlinearRegression(BaseModel):
             use_widgets=False,
             plot_fit=True,
             **kwargs):
-        # Get pobs for nonlinear fit.
-        if y is None:
-            y = self.gpm.phenotypes
-        # Else, numpy array or dataframe
-        elif type(y) == np.ndarray or type(y) == pd.Series:
-            pass
-        else:
-            raise FittingError("y is not valid. Must be one of the following: "
-                               "'obs', 'complete', numpy.array, pandas.Series."
-                               " Right now, its {}".format(type(y)))
-
         # Fit linear portion
         self._fit_additive(X=X, y=y)
 
@@ -183,7 +172,6 @@ class EpistasisNonlinearRegression(BaseModel):
 
     def _fit_additive(self, X=None, y=None, **kwargs):
 
-
         if hasattr(self, 'gpm') is False:
             raise Exception("This model will not work if a genotype-phenotype "
                             "map is not attached to the model class. Use the "
@@ -208,6 +196,7 @@ class EpistasisNonlinearRegression(BaseModel):
 
         return self
 
+    @arghandler
     def _fit_nonlinear(self, X=None, y=None, **kwargs):
         """Estimate the scale of multiple mutations in a genotype-phenotype
         map."""
@@ -236,7 +225,6 @@ class EpistasisNonlinearRegression(BaseModel):
             # Store items in case of error.
             nonlocal last_residual_set
             last_residual_set = (params, ymodel)
-
             return y - ymodel
 
         # Minimize the above residual function.
@@ -261,12 +249,9 @@ class EpistasisNonlinearRegression(BaseModel):
         # Point to nonlinear.
         self.parameters = self.Nonlinear.params
 
+    @arghandler
     def fit_transform(self, X=None, y=None, **kwargs):
-
         self.fit(X=X, y=y, **kwargs)
-
-        if y is None:
-            y = self.gpm.phenotypes
 
         linear_phenotypes = self.reverse(y, *self.parameters.values())
 
@@ -280,16 +265,12 @@ class EpistasisNonlinearRegression(BaseModel):
         return gpm
 
     def predict(self, X=None):
-
         x = self.Additive.predict(X=X)
         y = self.function(x, *self.parameters.values())
         return y
 
+    @arghandler
     def predict_transform(self, X=None, y=None):
-
-        if y is None:
-            y = self.gpm.phenotypes
-
         return self.function(y, *self.parameters.values())
 
     @arghandler
@@ -297,10 +278,6 @@ class EpistasisNonlinearRegression(BaseModel):
         # ----------------------------------------------------------------------
         # Part 0: Break up thetas
         # ----------------------------------------------------------------------
-        # Get thetas from model.
-        if thetas is None:
-            thetas = self.thetas
-
         i, j = len(self.parameters.valuesdict()), self.Additive.epistasis.n
         parameters = thetas[:i]
         epistasis = thetas[i:i + j]
@@ -313,68 +290,39 @@ class EpistasisNonlinearRegression(BaseModel):
 
         return ynonlin
 
+    @arghandler
     def hypothesis_transform(self, X=None, y=None, thetas=None):
-        # Get thetas from model.
-        if thetas is None:
-            thetas = self.thetas
-
         i, j = len(self.parameters.valuesdict()), self.Additive.epistasis.n
         parameters = thetas[:i]
-        epistasis = thetas[i:i + j]
+        epistasis = thetas[i:i+j]
 
         # Part 2: Nonlinear portion
-        return self.function(y, *parameters)
+        linear_phenotypes = self.reverse(y, *self.parameters.values())
 
+        # Transform map.
+        gpm = GenotypePhenotypeMap.read_dataframe(
+            dataframe=self.gpm.data,
+            wildtype=self.gpm.wildtype,
+            mutations=self.gpm.mutations
+        )
+        gpm.data['phenotypes'] = linear_phenotypes
+        return gpm
+
+    @arghandler
     def score(self, X=None, y=None):
-        # Get pobs for nonlinear fit.
-        if y is None:
-            pobs = self.gpm.phenotypes
-        # Else, numpy array or dataframe
-        elif type(y) == np.array or type(y) == pd.Series:
-            pobs = y
-
         xlin = self.Additive.predict(X=X)
         ypred = self.function(xlin, *self.parameters.values())
-        return pearson(pobs, ypred)**2
+        return pearson(y, ypred)**2
 
-
+    @arghandler
     def lnlike_of_data(self, X=None, y=None, yerr=None, thetas=None):
-        # ###### Prepare input #########
-        # If no model parameters are given, use the model fit.
-        if thetas is None:
-            thetas = self.thetas
-
-        # Handle y.
-        # Get pobs for nonlinear fit.
-        if y is None:
-            ydata = self.gpm.phenotypes
-        # Else, numpy array or dataframe
-        elif type(y) == np.array or type(y) == pd.Series:
-            ydata = y
-        else:
-            raise FittingError("y is not valid. Must be one of the following:"
-                               "'obs', 'complete', numpy.array, pandas.Series."
-                               " Right now, its {}".format(type(y)))
-
-        # Handle yerr.
-        # Check if yerr is string
-        if yerr is None:
-            yerr = self.gpm.std.upper
-
-        # Else, numpsy array or dataframe
-        elif type(y) != np.array and type(y) != pd.Series:
-            raise FittingError("yerr is not valid. Must be one of the "
-                               "following: 'obs', 'complete', numpy.array, "
-                               "pandas.Series. Right now, its "
-                               "{}".format(type(yerr)))
-
         # ###### Calculate likelihood #########
         # Calculate ymodel
         ymodel = self.hypothesis(X=X, thetas=thetas)
 
         # Likelihood of data given model
         return (- 0.5 * np.log(2 * np.pi * yerr**2) -
-                (0.5 * ((ydata - ymodel)**2 / yerr**2)))
+                (0.5 * ((y - ymodel)**2 / yerr**2)))
 
 
 class EpistasisNonlinearLasso(EpistasisNonlinearRegression):
@@ -441,41 +389,13 @@ class EpistasisNonlinearLasso(EpistasisNonlinearRegression):
             alpha=alpha,
             order=1, model_type=self.model_type)
 
+    @arghandler
     def lnlike_of_data(self, X=None, y=None, yerr=None, thetas=None):
-        # ###### Prepare input #########
-        # If no model parameters are given, use the model fit.
-        if thetas is None:
-            thetas = self.thetas
-
-        # Handle y.
-        # Get pobs for nonlinear fit.
-        if y is None:
-            ydata = self.gpm.phenotypes
-        # Else, numpy array or dataframe
-        elif type(y) == np.array or type(y) == pd.Series:
-            ydata = y
-        else:
-            raise FittingError("y is not valid. Must be one of the following:"
-                               "'obs', 'complete', numpy.array, pandas.Series."
-                               " Right now, its {}".format(type(y)))
-
-        # Handle yerr.
-        # Check if yerr is string
-        if yerr is None:
-            yerr = self.gpm.std.upper
-
-        # Else, numpsy array or dataframe
-        elif type(y) != np.array and type(y) != pd.Series:
-            raise FittingError("yerr is not valid. Must be one of the "
-                               "following: 'obs', 'complete', numpy.array, "
-                               "pandas.Series. Right now, its "
-                               "{}".format(type(yerr)))
-
         # ###### Calculate likelihood #########
         # Calculate ymodel
         ymodel = self.hypothesis(X=X, thetas=thetas)
 
         # Likelihood of data given model
         return (- 0.5 * np.log(2 * np.pi * yerr**2) -
-                (0.5 * ((ydata - ymodel)**2 / yerr**2)) -
+                (0.5 * ((y - ymodel)**2 / yerr**2)) -
                 (self.alpha * sum(abs(thetas))))
